Make attempt at StorageIoInterface info pre Android 15
Older versions of Android are not assured to have the "/tmp"
directory.  Previously, we simply skipped collecting this info
for any device with a first API before Android 15.

However, this info is useful on older devices as well.  So we
at least try a couple tmp directories on the device before
giving up on collecting info.

Bug: 432823222
Test: Using "atest android.edi.cts.StorageIoInterfaceDeviceInfo", we passed on device after Android 15, collecting info; also passed on device before Android 15, collecting info; changing DEVICE_TMP_DIRS to bad paths we passed, collecting no information.
Flag: EXEMPT for collecting unflagged info
Change-Id: I4ac9881b68b8ae285d86a84e5449de0299edf996