Add CTS for BarringInfo#getCellIdentity
This CL adds CTS tests to verify the end-to-end functionality of the BarringInfo#getCellIdentity() API.

The tests are added for both TelephonyCallback and the legacy PhoneStateListener to ensure complete coverage for clients. The new tests register a listener for onBarringInfoChanged and, upon receiving a callback, validate that the returned CellIdentity is non-null and an instance of a known, valid subclass (e.g., CellIdentityLte, CellIdentityNr).

These tests are gated by the cell_identity_for_barring_info feature flag and verify the functionality exposed in a companion CL.

Bug: 431881770
Bug: 412452528
Test: atest BarringInfoTest
Test: atest TelephonyCallbackTest, atest PhoneStateListenerTest
Flag: com.android.internal.telephony.flags.cell_identity_for_barring_info

Change-Id: I64223b1cd811ef59a1eb4c50d94ec62f5e23057b