Increase resilience for the SteadyTimestampCodec
MediaCodecs are not as fast on all devices, so increase the resilience
of the test, adding ability to recover for when a frame is skipped and
the VirtualCameraThread duplicates a frame and increases the timestamp.

Also, improved logging and how it can be expanded at runtime using "adb
shell setprop log.tag.SteadyTimestampCodec DEBUG", though by default
it's only logged above warn.

Bug: 419598631
Bug: 428855045
Test: atest android.virtualdevice.cts.camera.VirtualCameraCaptureTest#captureMultipleImages_withCustomTimestamp_withMediaCodec --iterations 10
Flag: EXEMPT bugfix
Change-Id: Ia884690fe11faf26f59a93a9f61bf81f512414d4