Fix FullBackupOnlyHostSideTest for multi-user.
The test was failing on HSUM devices where the main user is not user 0.
Backup command was not specifying the user, defaulting to user 0.
This caused assertion failures as the test app (running as user 10) could not see the restored data.
Explicitly pass the current user ID to all package install calls.

Flag: TEST_ONLY
Test: atest FullBackupOnlyHostSideTest
Bug: 413811840
Bug: 414774137
Bug: 413021217
Change-Id: I6e13ef04511545df67f28eb983cce6015c9e698d