Fix flaky streamExpansion tests in OutputSwitcherTest
Some helper methods in this test class look for a node in the tree
with a given route name, and then search to find a nearby element in
the UI tree with accessibility text indicating it's the thing to click
to either add a route to, or remove a route from, the ongoing
stream. This code was written assuming the node to click appears in
the tree before or at the same time as the route name node, but
sometimes that doesn't happen and the test fails. The fix is to wait a
little while for the node with accessibility text to show up.

Bug: 435257225
Test: atest OutputSwitcherTest
Flag: EXEMPT fixing a flaky test
Change-Id: I124c0fd7ad4d9c3eaedf386c14b1478a635d4015