Add a test for bringup of multiple cell networks
Add a test that evaluates multiple cellular networks,
comparing them against each other to ensure that the
networks that are brought up are valid. There are
some limitations on this test, which prevent it from
being comprehensive:
1) We do not currently require that the cell service used
   for CTS testing is configured with multiple PDNs.
2) We do not currently require that the network used for
   CTS testing supports/requires multiple concurrent PDNs.

As such, for now the test can attempt to bring up any and
all cellular networks that are available, and for each
one, it can compare the link properties against all of the
others, ensuring they aren't aliasing to the same
underlying Linux interface.

Bug: 396730178
Test: atest DataConnectionTest
Test: verified on aosp_cuttlefish-userdebug and panther-eng
Flag: EXEMPT test-only change

Change-Id: Ib0f9e219f98300414bd42fedd01a0577a88d352c