WebcamTestActivity: Skip if UVC is not supported
Currently, CtsVerifier relies on the host script to notify it when
UVC Gadget mode is not supported. This is not ideal as it requires
vendors to set up Webcam test, even if their device doesn't support
if.

With I12fb1c003415f, there is now a Test API that allows CtsVerifier
to query UVC Gadget support without asking the host script.

This CL uses the test API to skip the test early and removes all logic
around the host script reporting uvc support.

Bug: 434155777
Test: Verified that CtsVerifier can still fetch UVC support status
Flag: EXEMPT. Test only change.
Change-Id: Ib124b755be55edf03fe7faf6e996e6bf553fc822