Fix testDialogPositionChangedAfterImeIsShown for Automotive multi-window
In multi-window surfaces (like Automotive multi-window), the default
mock IME height is not large enough to overlap with the test's dialog
window (activity may show up on the top and mock IME on the absolute
bottom without any overlapping).

This change addresses this by setting the mock IME height to 50% of the
test activity's area when running on Android Automotive with
multi-window, regardless of the test activity's position.

Test: atest CtsInputMethodTestCases:KeyboardVisibilityControlTest#testDialogPositionChangedAfterImeIsShown
Fix: 434892869
Flag: TEST_ONLY
Change-Id: I0a75c569edf2698088b07c88716c2e66025b6086