Set fixed times on MotionEvents injected by BackGestureTouchHelper
This CL removes the UinputTouchScreen from BackGestureTouchHelper.
UinputTouchScreen does not allow setting the MotionEvent timestamps.
However it is necessary to control the timestamps of the injected
MotionEvents to reduce test flakiness.

This fixes a bug that back gestures were cancelled due to longpress
detection in EdgeBackGestureHandler. This happened when the main thread
was very busy during tests and therefore MotionEvents were delayed.

Bug: 436563664
Bug: 425652879
Test: BackGestureInvokedTest
Flag: EXEMPT test bugfix
Change-Id: I66aa3c5451444072c182cd3afee3984df65e315e