Refactor CheckFlagsRule for tests with `@RequiresFlags[Enabled|Disabled]`
To enable Global lint CheckFlagsRuleDetector, we need to first fix
(or baseline) all existing errors. The CheckFlagsRuleDetector lists all
fields of an instance so need to make CheckFlagsRule a field to avoid
lint errors. No logic changes.

For context on the development of this new lint, see b/430031210.

Bug: 435371380
Test: m .
Flag: EXEMPT test change
Change-Id: I6e51ac4e9eb16d54f19d115a571f31389195281a