Relax ProcessTest.testMiscMethods
setThreadPriority specifies its second argument with

@IntRange(from = -20, to = THREAD_PRIORITY_LOWEST)

The documentation comment says

@param priority A Linux priority a.k.a. "niceness" level, from -20 for
highest scheduling priority to 19 for lowest scheduling priority.

We were testing for specific behavior with an input of -21. My
interpretation of both of the above is that this behavior is explicitly
unspecified, so throwing an exception is not only fine, but better
behavior than reinterpreting the argument as -20. So have the test
require an IllegalArgumentException instead to improve debuggability.

ag/34421756 adjusts the implementation to satisfy this requirement.

Bug: 417492941
Flag: EXEMPT TEST_ONLY
Test: atest ProcessTest, choosing the modified test
Change-Id: Ib60e71990468a0371d3ff39a561721061e6a03eb