Adjust WindowMetricsActivityTests freeform bounds to avoid top inset
The freeform bounds used in WindowMetricsActivityTests can overlap
with the top display inset on some devices. As a result, the bounds are
shifted downwards to avoid this clash. However, the test does not expect
this bound shift and instead expects the freeform bounds that are set.

Adjust the fixed `WINDOW_BOUNDS` to be further down and thus avoid any
possible display insets.

Flag: EXEMPT test fix
Bug: 436229724
Test: atest CtsWindowManagerDeviceWindow:WindowMetricsActivityTests
Change-Id: I495c5c9beff741be34db6eeb66cf53b434484fa1