Force WebcamTestActivity to pass until API 36.1
The WebcamTestActivity was broken and passed unconditionally until 25Q2.
To prevent failures in minor versions, the test is now forced to pass
unconditionally for API levels up to and including Baklava (API 36). A
Toast message is shown when this occurs, indicating that the test will
likely fail starting from API 36.1.

Bug: 437352393
Test: Verified that the test always passes on SDK <= 36
Flag: EXEMPT. Test only.
Change-Id: I54b2239da0a32932d160077ea9a7735611347a2d