Fix WINDOWING_MODE_FULLSCREEN usages in KeyboardVisibilityControlTest
This fixes the usages of launching the test app in fullscreen in
KeyboardVisibilityControlTest.

If the test activity needs to be fullscreen, this change adds
asNewTask() to the start option. Without it, the activity can be
started in the same stack of StateInitializeActivity, which may not be
in fullscreen.

testImeVisibilityWhenDismissingDialogWithImeFocused:
  The activity must be in fullscreen to get IME inset control.

runImeDoesntReshowAfterKeyguardTest:
  Making the activity fullscreen is not needed.

testDialogPositionChangedAfterImeIsShown:
  In order to verify that the opend dialog position being adjusted by
  IME, the test activity itself should be fullscreen so that activity
  itself has a stable bounds while the dialog can move around in the
  screen.

runRestoreImeVisibility:
  Making the activity fullscreen is not needed.

There are a few more usages of WINDOWING_MODE_FULLSCREEN. They require
non-trivial changes and will be addressed in separate changes.

Bug: 438635422
Test: KeyboardVisibilityControlTest
Flag: EXEMPT text fix
Change-Id: I21893f7a2029ad4a9d4ea46d5210d41f13920f62