[ContentCaptureCTS] Modify ContentCaptureTests to use current CTSContentCaptureServiceInstance to avoid flakiness.
In the tests depending on the ContentCaptureService lifecycle, caching
the instance of CtsContentCaptureService has been causing falkiness
since in case of a reconnect/recreation of service the older instance in
cached in the testcase and the expected service sessions are not found
when queried. This fix tries to get the latest instance of the service
from the service watcher object and then use it to assert the session
values and test cases.

Bug: 418182321

Test: atest AppKilledTest.java
Test: atest BlankActivityTest.java
Test: atest ChildlessActivityTest.java
Test: atest CutomViewActivityTest.java
Test: atest LoginActivityTest.java
Test: atest PartialNotImportantActivityTest.java
Test: atest ResizingEditActivityTest.java
Test: atest BlankWithTitleActivityTest.java

Flag: EXEMPT, tests changes only.
Change-Id: Id518bdfabb50508c14247ed1e089162a2a22bdc8