mediav2: mark av1 params as CODEC_ANY instead of CODEC_ALL
The intent of CodecEncoderTest is to verify all encoder components for
various media codec api. The test is picking low resolution encoding
configurations during testing with the idea that components on slow-cpu
devices, components with very limited capabilities will also be covered.
However due to reasons in b/436122742, this may not always hold true.

The best way to address this is pick configuration at run time based on
the advertised capabilities. This will be done in future CLs. For now
bypass the failure seen.

Bug: 436122742
Test: atest CtsMediaV2TestCases:CodecEncoderTest
Test: atest MctsMediaV2TestCases:CodecEncoderTest
Flag: TEST_ONLY
Change-Id: I3b8f5f5b5651e3fb87835a23fce31365796ad0de