[WV] Add new CTS module for webkit multiuser tests
We were testing WebView's behaviour on multiuser
scenarios using hostsidetests. As a result, we managed Android user
profile creation within these tests. This was complicated and
flaky.

This changes introduces the go/bedstead library which handles
multiuser testing without us needing to write explicit code for
it. Unfortunately, we need to add this as a new module since
Bedstead doesn't work correctly with test ochestrator that is
enabled in our existing on-device CTS test library. Additionally,
the test setup time is increased when using bedstead even if
we run a non-multiuser test so keeping it as a separate module
is preferable.

The hostside multiprofile test is deleted.

Flag: TEST_ONLY
Bug: 381862299
Test: Locally ran CTS tests
Change-Id: I69f57ac2c19da7005d848d9b94fa7d42d79a560c