Change config to use CtsWebkitMultiuserTestCases
This new module was introducted to support multiuser testing
of webkit APIs on the device side without needing a hostside
test.
The test cases in this module were previously part of
CtsHostsideWebViewTests module. Therefore, I am adding this
module to config options that refers to CtsHostsideWebViewTests
whereever it makes sense.

Test: Ran CTS locally
Flag: TEST_ONLY
Bug: 381862299
Change-Id: I04ebb2e3c3870c8c6b7b832b2f7e146479194da3