Propagate callback assertions in MockWebViewClient
The implementation of `onPageFinished` asserts that the object has
previously received other callbacks. This CL ensures that the assertion
errors are propagated to the instrumentation thread if the thread is
currently waiting on a loadUrl call. This improves the quality of the
resulting stack traces by highlighting which call to
loadUrlAndWaitForCompletion actually triggered the assertion error.

If the instrumentation thread is not waiting for the loadUrl operation
to complete, then the exceptions are simply thrown on the main thread as
before.

Bug: 440523702
Test: atest CtsWebkitTestCases
Flag: TEST_ONLY
Change-Id: Ifa45246de04f99e4e2ac822b505a13e1044d0c74