Install backup test APKs for the current user only.
This fixes PermissionTest#testDelayedRestore, which fails on HSUM devices where the main user is not user 0 (e.g., user 10).

The root cause is that the test APK is installed for all users by default.
When the test runs for a specific user (e.g., user 10), the package manager treats this as an update rather than a new installation.

This causes the doRestore flag in InstallPackageHelper#restoreAndPostInstall to remain false, which in turn prevents the delayed permission restore logic from being triggered.
As this restore logic is the subject of the test, the test fails.

By adding the `--user current` install argument to the test configuration, this change ensures that pm install installs the test package only for the current user under test.
The installation is now correctly identified as a new install, allowing the permission restore logic to execute and the test to pass.

Bug: 415449393
Bug: 440976861
Test: atest android.backup.cts.PermissionTest
Test: atest CtsBackupTestCases
Flag: TEST_ONLY
Change-Id: I54b92f54507a3acf7ec834cd5948866e3e4719b3