Enable Winscope traces for ConcurrentMultUserImeTest
Enabled winscope tracing using the instructions in
go/if-doc/winscope/tracing-in-tests.md

The device-listeners option ntegrates the Winscope trace collector
into the test run, as described in
go/if-doc/winscope/tracing-in-tests.md

The FilePullerLogCollector is configured to pull the generated
trace files (identified by perfetto_file_path), making them
available for analysis.

The addition of the ScreenshotOnFailureCollector provides a
screenshot of the device state when a test fails.

Bug: 439450593
Flag: TEST_ONLY
Test: atest ConcurrentMultiUserTest
Change-Id: I6970dd75b2e6f5d08c79b2bf7aef8b0f548be997