Fix flake in testSetPictureListener
I suspect that the flake is caused by the WebView field being read
before it has been set, since the blocking queue is updated before the
field is written.

This CL reordes the operations to ensure that the mWebView field is
written before the queue is updated and the test stops waiting.

Bug: 440709413
Bug: 440693569
Test: atest CtsWebkitTestCases
Flag: TEST_ONLY
Change-Id: Ib91faf2022fa894481d38ef27c1586a7abaf9631