Fix flaky requestIari_WithReadPrivilegedPermission
The requestIari_WithReadPrivilegedPermission method was sometimes throwing a security exception because it returned the required permissions before the processing of the method.
To resolve this issue, we've modified the test to retain the required permissions until the result callback is fully executed. This ensures that the permissions are available throughout the test lifecycle, allowing it to complete as intended without encountering the security exception.

Bug: 440173302
Bug: 441555787
Test: atest --rerun-until-failure 100 -c android.telephony.cts.TelephonyManagerTest#requestIari_WithReadPrivilegedPermission
Flag: TEST_ONLY
Change-Id: I2a07fc3fae1e01d82bead8df056ca77ca3a3f3de