CtsSurfaceControlTests: add input-related subset
The inputflinger TEST_MAPPING currently includes the whole of
CtsSurfaceControlTests, even though most of the tests don't depend on
the input subsystem. Add a config that only runs the subset that are
input-related (based on me skimming the test case names in each test
file). This should prevent unrelated failures from being filed to the
input team.

Test: $ atest CtsSurfaceControlTests_input_related
Bug: 440698094
Flag: TEST_ONLY
Change-Id: I057cd362b7393d725267998304900c045890054d