Refactor monet.Style to ThemeStyle
Deprecates `monet.Style` in favor of the new `ThemeStyle` class in preparation for the upcoming `ThemeManagerService`.

This change is purely a refactoring, moving the existing `monet.Style` class to a new location and name with no functional changes. All existing usages in SystemUI have been updated to the new class.

Bug: 441912053
Test: Existing tests updated.
Flag: EXEMPT mechanical refactoring
Change-Id: I8950bd5392adb2d2b9cf53d6325d8ea0017de823