Fix NPE in dpmwrapper CTS tests
http://ag/35390209 is causing some NPE in CTS tests. It removed the
try-catch present in the Java file since the comment -
"Should never happen, but needs to be catch as some methods declare
checked exceptions" seemed to indicate that the try-catch was for
catching checked exceptions and not real failures. Since none of the
mocked / spied methods had checked exceptions, I assumed that it was
safe to remove.

Turns out that a bunch of CTS tests are actually incorrect. They are
providing null values to places which don't expect null values, and
these NPEs are getting blanket caught by this try-catch.

This CL is a stopgap to make failing CTS tess pass while a real fix is
worked upon. Please see the attached bug for more context.

Bug: 442697334
Test: atest com.google.android.cts.settings.PrivacyDeviceOwnerTest
Flag: TEST_ONLY
Change-Id: I55e9d1e52684ad20987b73ae4f31074efbcaa8e7