CtsWifiLocationTestApp: Add CHANGE_NETWORK_STATE
Desktop devices in the lab are tested using an Ethernet connection. When
an Ethernet connection is present, WiFi is no longer the default network
and therefore does not have the FOREGROUND capability. This causes the
test to fail to find a match for the WiFi network.

This change adds the CHANGE_NETWORK_STATE permission. This allows the
app to match WiFi networks even if they lack the FOREGROUND capability,
enabling the tests to pass.

Bug: 435264187
Test: Ran the two test modules
WifiLocationInfoBackgroundTest/WifiLocationInfoForegroundTest on Pixel
and Desktop [[ PASS ]]
Flag: TEST_ONLY

Change-Id: Idd69eb9021917632b8134f03b40645f2ef3466c1