Wait for animations to start in parallel
Before this CL, the code was waiting for the animations to start in two
places, in series - first, via a condition variable, and next, using
'wmState'. This is problematic because the animation could be over after
the first wait was completed, so the second wait would just fail.

Instead, in this CL, we make the two waits run in parallel. Besides
improving the test speed, this could also reduce the flake rate because
the device is going to be more likely in a state where the animation is
still running (which is what is desired for these tests).

Bug: 443038461
Test: atest WindowUntrustedTouchTest
Flag: EXEMPT TEST_ONLY
Change-Id: Ied11b1433ab1d1f2fef1ce5bbf6401db100cd0f2