Check for exceptions in the other thread before continuing
Before this CL, we checked for the result of the future at the end of
the test. However, the future might fail (polling check could throw an
exception). If that happens, then this would only be surfaces at the end
of the test run.

However, if the test was already failing (system was in a bad state),
then it's likely that *another* polling check would also fail. This
means that we wouldn't ever get to see the earlier failure. This would
make it more difficult to debug this test.

The only reason this was found is because the logs contained *another*
exception, where the receiver was never unregistered, even though the
code for unregistering the receiver was present. This will be fixed in a
separate CL.

For this CL, though, we will just move the check for exceptions before
continuing the test. This will help surface these failures sooner.

Bug: 443882075
Test: atest IncompleteMotionTest
Flag: EXEMPT TEST_ONLY
Change-Id: I64e780da82a896cef57cf43a5e3c8c4fdf91186b