Improve helpers that read flags.
They could crash the test if the flag cannot be read, even if the
test doesn't use it.

Ideally such crashes shouldn't happen, but we can take the "defensive
approach" or treating the flag as disabled - if the test depends on
the flag being enabled, it would fail regardless.

Test: atest DeviceOwnerTest#testDisallowFactoryReset
Bug: 442333334
Bug: 383180991
Flag: TEST_ONLY

Change-Id: I72e23911f62291a9a72f45f729ff1873493a9504